Refactor health check and fix unit tests
